+++ packages/url_launcher/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.1
 
 * Support option to launch default browser in iOS.
+* Parse incoming url and decide on what to open based on scheme.
+* Support WebView on Android.
 
 ## 1.0.0